Avoid circular dependency of util and machine
