+++ integration/experiment/machine.py
@@ -91,3 +91,18 @@ class Machine:
         for sn in signal_names:
             self.signals[sn] = util.geopmread('{} board 0'.format(sn))
 
+def init_output_dir(output_dir):
+    if not os.path.exists(output_dir):
+        os.makedirs(output_dir)
+    elif not os.path.isdir(output_dir):
+        raise RuntimeError('Requested output directory is a file: {}'.format(output_dir))
+    mm = Machine(output_dir)
+    mm.save()
+    return mm
+
+
+def get_machine(output_dir):
+    mm = Machine(output_dir)
+    mm.load()
+    return mm
+
